Did lots of GUI stuff
Got the "Planner Panel" mostly looking as it should, it has the planner
and shows the ingredients for each meal in the planner when a day is
selected. Started working on a CookBook Panel and added skeleton sort of
for a MealCreatorPanel where the user can add meals to the cook book,
and from there add them to the planner.